+++ CHANGELOG.md
@@ -17,6 +17,10 @@
 
     In TypeScript, `export {Type}` is supposed to be silently removed by the compiler if `Type` does not refer to a value declared locally in the file. Previously this behavior was incompletely implemented. The statement itself was removed but the export record was not, so later stages of the pipeline could sometimes add the export statement back. This release removes the export record as well as the statement so it should stay removed in all cases.
 
+* Forbid exporting non-local symbols in JavaScript
+
+    It is now an error to export an identifier using `export {foo}` if `foo` is not declared locally in the same file. This error matches the error that would happen at run-time if the code were to be evaluated in a JavaScript environment that supports ES6 module syntax. This is only an error in JavaScript. In TypeScript, the missing identifier is silently removed instead since it's assumed to be a type name.
+
 ## 0.6.33
 
 * Fix precedence of tagged template expressions ([#372](https://github.com/evanw/esbuild/issues/372))